Renaming route and schema files as per new teams convention #patch (#46)
https://eaflood.atlassian.net/browse/SFDP-554

As part of our ongoing effort to refactor the front-end codebase, this PR includes bulk file renames.
Following our new team coding conventions, all route, schema, services and presenter file names now include a file-type suffix. For example:

OLD: /routes/home.js
NEW: /routes/home-routes.js

OLD: /schemas/business/telephone.js
NEW: /schemas/business/telephone-schema.js

The purpose of this change is to keep the file names in line with the top file function, which for routes, schemas, presenters and services should include the file type in the function name.
