+++ test/unit/routes/business/business-legal-status-check-routes.test.js
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'vitest'
-import { businessLegalStatusRoutes } from '../../../../src/routes/business/business-legal-status-change.routes.js'
+import { businessLegalStatusRoutes } from '../../../../src/routes/business/business-legal-status-change-routes.js'
 const [getBusinessLegalStatusChange] = businessLegalStatusRoutes
 
 describe('change business legal status', () => {
